[ARM] tegra: nvrm: duplicate all handles on AVP pin
one of the AVP modules used by camera apparently pins a handle
without duplicating the handle onto the AVP first. this is
an incorrect usage of the API, but was only flagged as an error
with the new nvmap implementation.

work-around this for now by performing an explicit duplicate on
pin, and an explicit free on unpin.

Change-Id: I49777fdd1009e564ac326fda00b96275d0ae5c19
Signed-off-by: Gary King <gking@nvidia.com>