[bt/rfkill]: make either RST/SHUTDOWN GPIO usage optional.
the current BCM4329 rfkill driver assumes usage of 2 GPIOs
known as RST and SHUTDOWN and the driver makes a particular
GPIO mandatory. Some of the platforms does not define both
GPIOs, instead a single either RST/SHUTDOWN GPIO is used to
setup the chip.

This change makes driver to consider any of the two GPIOs
as optional and use any of the RST/SHUTDOWN GPIOs.

Simultaneous usage of both GPIOs is also allowed.

Change-Id: Ib66ea350e78642082f639514ef7a9def6e460e28
Reviewed-on: http://git-master/r/14534
Reviewed-by: Anantha Idapalapati <aidapalapati@nvidia.com>
Tested-by: Anantha Idapalapati <aidapalapati@nvidia.com>
Reviewed-by: Bharat Nihalani <bnihalani@nvidia.com>