ARM: tegra: bt-rfkill: remove usage of reset GPIO.
The current BCM4329 rfkill driver assumes usage of 2 GPIOs
known as RST and SHUTDOWN. Some of the platforms do not
define both GPIOs, instead a single SHUTDOWN GPIO is used to
setup the chip.

This change makes driver to consider the RST GPIO as optional
and use the RST GPIO only when it is defined as a platform resource.

As ventana doesn't use RST GPIO, removing its definition.

Change-Id: Iae87d2d5093ff0390c7e7ddca5b95b1b44afd55d
Reviewed-on: http://git-master/r/14535
Reviewed-by: Bharat Nihalani <bnihalani@nvidia.com>
Tested-by: Bharat Nihalani <bnihalani@nvidia.com>