serial: core:  skip call set_termios/console_start when no_console_suspend
The commit 4547be7 rewrites suspend and resume functions, this
introduces a problem on the OMAP3EVM platoform. when the kernel boots
with no_console_suspend and we suspend the kernel, then resume it,
the serial console will be not usable. This problem should be common
for all platforms.
The cause for this problem is that when enter suspend, if we choose
no_console_suspend, the console_stop will be skiped. But in resume
function, the console port will be set to uninitialized state by
calling set_termios function and the console_start is called without
checking whether the no_console_suspend is set, Now fix it.

Refer
http://git.kernel.org/?p=linux/kernel/git/stable/linux-2.6.37.y.git;
a=commit;h=ca2e71aa8cfb0056ce720f3fd53f59f5fac4a3e1 for

cherry-pick the changes from the kernel 2.6.37, commit id as
ca2e71aa8cfb0056ce720f3fd53f59f5fac4a3e1

Bug 758845

Change-Id: Ib83494b2b13cd7179d3ae173be328c8cd776d431
Signed-off-by: puneet saxena <puneets@nvidia.com>
Reviewed-on: http://git-master/r/19590
Reviewed-by: Bharat Nihalani <bnihalani@nvidia.com>
Tested-by: Bharat Nihalani <bnihalani@nvidia.com>
Reviewed-by: Varun Wadekar <vwadekar@nvidia.com>