tegra: sdhci: enable write protection on sd cards
Write protection was disabled in K36. Enabling this feature by
storing the write protection GPIO pin number (wp_gpio) from
platform_data for the device and reading this GPIO (whenever the
card is inserted) to determine whether the card is write protected
or not.
Bug 686892.

Change-Id: I6f1855b7b974caaa5ea58aaee3881d85d5da6c18
Reviewed-in: http://git-master/r/#change,21101
Reviewed-on: http://git-master/r/23971
Reviewed-by: David Schalig <dschalig@nvidia.com>
Tested-by: David Schalig <dschalig@nvidia.com>
Reviewed-by: Pavan Kunapuli <pkunapuli@nvidia.com>
Reviewed-by: Bharat Nihalani <bnihalani@nvidia.com>