video: pwm_backlight: Add check_fb hook
In systems with multiple framebuffer devices, one of the devices might
be blanked while another is unblanked.  In order for the backlight
blanking logic to know whether to turn off the backlight for a
particular framebuffer's blanking notification, it needs to be able to
check if a given framebuffer device corresponds to the backlight.

This plumbs the check_fb hook from core backlight through the
pwm_backlight helper to allow platform code to plug in a check_fb hook.

Originally reviewed on http://git-master/r/21716

Reviewed-on: http://git-master/r/23100
(cherry picked from commit c7831cb27fc6a2701b020676a6daf9c76e9470d9)

Change-Id: I32895d6f2c5988b09680a350e169f1f3c9b56da6
Reviewed-on: http://git-master/r/23725
Reviewed-by: Robert Morell <rmorell@nvidia.com>
Tested-by: Gaurav Sarode <gsarode@nvidia.com>
Reviewed-by: Mayuresh Kulkarni <mkulkarni@nvidia.com>
Reviewed-by: Bharat Nihalani <bnihalani@nvidia.com>