ASOC: tegra: Restore jack switch state after Suspend
Adds a function to restore jack state after suspend, it will notify
the upper layer about the current states of the jack after suspend.

In suspend state the notifier function doesn't get triggered, leaving the
system in an unknown state. This function will be called on resume after
suspend.

Fix for Bug: 804328

Change-Id: I89ab32a6d775904fb1a71725d4e1fbda89002431
Reviewed-on: http://git-master/r/26853
Reviewed-by: Sumit Bhattacharya <sumitb@nvidia.com>
Tested-by: Sumit Bhattacharya <sumitb@nvidia.com>
Reviewed-by: Boris Suvorov <bsuvorov@nvidia.com>
Tested-by: Boris Suvorov <bsuvorov@nvidia.com>
Reviewed-by: Varun Colbert <vcolbert@nvidia.com>